changed file

+++ clogrot.sh
@@ -2,15 +2,17 @@
 clear
 echo "----INACTIVE LOG VERIFIER-----"
 echo
-echo "this script get all 'expired files' and compress then in single file."
-echo "A expired file is a archive that not was accessed neither modified 30 days ago"
+echo "this script get all 'expired files' and compress then in single file, removing the old ones."
+echo "A expired file is a archive that not was accessed neither modified 60 days ago"
 echo
 echo
 
 
-time_file_inactive="30 days ago";
+time_file_inactive="60 days ago";
 date_expired_files=`date  +%Y-%m-%d  --date="$time_file_inactive"`;
 
+
+#echo $date_expired_files
 
 #echo "All files not modified or accessed since "$date_expired_files \n
 #"will be compacted in a zip file and removed from log directory";
@@ -23,6 +25,7 @@ default_tar_file_name="compressed_logs_"`date  +%Y-%m-%d`".tar";
 count_rm_file=0;
 c=0;
 splitSeparator="|";
+wasFoundOldFile = false;
 
 
 while [[ $c != ${#array_all_files[*]} ]]; do
@@ -41,35 +44,57 @@ while [[ $c != ${#array_all_files[*]} ]]; do
         #echo $file_name;
         #echo $last_access;
         #echo $last_modification;
-        #echo "";
+	#echo $date_expired_files
+        #echo "/////////////";
 	
 	
         if [[ $(date -d $last_modification +%s) -le $(date -d $date_expired_files +%s) || $(date -d $last_access +%s) -le $(date -d $date_expired_files +%s) ]]
         then
-    	   array_files_to_remove[$count_rm_file]=$file_name;
-	   (( count_rm_file++ ));     
+
+		
+		echo $(date -d $last_modification +%s)
+		echo "le"
+		echo $(date -d $date_expired_files +%s)
+		echo ""
+
+		echo "||"
+
+		echo $(date -d $last_access +%s)
+		echo "le"
+		echo $(date -d $date_expired_files +%s)
+		echo ""
+		wasFoundOldFile = true;
+
+		array_files_to_remove[$count_rm_file]=$file_name;
+		(( count_rm_file++ ));     
 	fi
 
         (( c++ ))
 done
 
-echo "files picked, wait for compression...";
+if [[ wasFoundOldFile -eq true ]]
+then
+
+	echo "files picked, wait for compression...";
 
 
-echo ""
-echo ""
-echo "reporting files compacted and removed:"
-c=0;
+	echo ""
+	echo ""
+	echo "reporting files compacted and removed:"
+	c=0;
 
-echo "the output filename is: " $default_tar_file_name;
+	echo "the output filename is: " $default_tar_file_name;
 
-
-while [[ $c != ${#array_files_to_remove[*]} ]]; do
-    #echo ${array_all_files[$c]};
-    tar zcvf $default_tar_file_name ${array_all_files[$c]}
-    rm -i ${array_all_files[$c]};
-    (( c++ ))
-done
-
+ 
+	while [[ $c != ${#array_files_to_remove[*]} ]]; do
+	    #echo ${array_all_files[$c]};
+	    tar zcvf $default_tar_file_name ${array_all_files[$c]}
+	    rm -i ${array_all_files[$c]};
+	    (( c++ ))
+	done
+else
+	echo "Theres no files to remove"
+	exit 0
+fi
